fix(Migration V10): Correção da variável valor_produto / fix(DialogConfirm): alterado o window.confirm() das telas de EmpresaTodos e Produto, ProdutoAlterar para um Dialog Confirm personalizado.

+++ cardapio-qr-code/src/main/resources/db/migration/V10__Create_Table_Produto.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS `produto` (
   	`id` bigint(20) NOT NULL AUTO_INCREMENT,
   	`produto_nome` varchar(255) DEFAULT NULL,         
-  	`valor_produto` decimal DEFAULT NULL,  
+  	`valor_produto` double DEFAULT NULL,
    `data_cadastro` date DEFAULT NULL,
 	`data_atualizacao` date DEFAULT NULL,
 	`disponivel` bit(1) DEFAULT NULL,
